pyproject.toml
[project]
name = "django-saml2-auth-community"
# Embedded as the PyPI project description (long description); required for pypi.org project page.
readme = { file = "README.md", content-type = "text/markdown" }
dynamic = ["version"]
description = "Django SAML2 Authentication Made Easy."
license = "Apache-2.0"
authors = [{ name = "Mostafa Moradian", email = "mstfmoradian@gmail.com" }]
requires-python = ">=3.10"
keywords = ["authentication", "django", "saml", "saml2", "sso"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = ["pyjwt==2.12.1", "pysaml2==7.5.4"]

[project.urls]
Homepage = "https://github.com/mostafa/django-saml2-auth"
Issues = "https://github.com/mostafa/django-saml2-auth/issues"

[dependency-groups]
dev = [
    "coverage==7.13.5",
    "cyclonedx-bom==7.3.0",
    "django-stubs==6.0.2",
    "httpx>=0.27,<1",
    "mypy==1.20.0",
    "pytest==9.0.3",
    "pytest-django==4.12.0",
    "responses==0.26.0",
    "ruff==0.15.9",
    "testcontainers==4.10.0",
    "types-pysaml2==1.0.1",
    "types-setuptools==82.0.0.20260408",
]

[build-system]
requires = ["hatchling==1.29.0", "uv-dynamic-versioning==0.14.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["django_saml2_auth"]

[tool.ruff]
exclude = [
    "dist",
    "build",
    "env",
    "venv",
    ".env",
    ".venv",
    ".tox",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
    ".ruff",
]
line-length = 100

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_saml2_auth.tests.settings"
pythonpath = "."
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::jwt.warnings.InsecureKeyLengthWarning",
    # pysaml2 warns when SP allows unsigned assertions; integration IdP uses unsigned responses.
    "ignore:The SAML service provider accepts unsigned SAML Responses and Assertions:UserWarning:saml2.client_base",
]
addopts = ["--import-mode=importlib"]
testpaths = ["django_saml2_auth/tests"]
markers = [
    "integration: tests that require Docker (SAML IdP container)",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "django_saml2_auth.tests.settings"
